# do not split surrogates in `ParameterInstance.ToDisplayText()`