Fixes #1703 

The original idea was to explicitly specify `UTF-8` as a `StreamWriter` encoding. It would cause `StreamWriter` to **not** throw an exception when only part of a surrogate pair left (it would be replaced with the replacement character).   

But actually updating `.ToDisplayText()` forbidding surrogate pairs splitting looks like a better fix.

Old `DisplayText`: `a😀☐(...)😀😀 [201]` (causes `StreamWriter` to throw the exception)
New `DisplayText`: `a😀(...)😀😀 [201]`

Logs:
```
// ***** BenchmarkRunner: Start   *****
// ***** Found 1 benchmark(s) in total *****
// ***** Building 1 exe(s) in Parallel: Start   *****
BuildScript: C:\Users\onovak\Documents\repos\BenchmarkDotNet\samples\BenchmarkDotNet.Samples\bin\Debug\net461\b681758f-a705-4498-bd08-6073e1006811.bat
// ***** Done, took 00:00:02 (2.91 sec)   *****
// Found 1 benchmarks:
//   PerfTests.OldBehavior: DefaultJob [_utf16String=a😀(...)😀a [202]]

// **************************
// Benchmark: PerfTests.OldBehavior: DefaultJob [_utf16String=a😀(...)😀a [202]]
// *** Execute ***
// Launch: 1 / 1
// Execute: C:\Users\onovak\Documents\repos\BenchmarkDotNet\samples\BenchmarkDotNet.Samples\bin\Debug\net461\b681758f-a705-4498-bd08-6073e1006811.exe --benchmarkName "BenchmarkDotNet.Samples.PerfTests.OldBehavior(_utf16String: "a😀😀😀😀😀😀😀😀😀😀😀😀😀😀😀😀😀😀😀😀😀😀😀😀😀😀😀😀😀😀😀😀😀😀😀😀😀😀😀😀😀😀😀😀😀😀😀😀😀😀😀😀😀😀😀😀😀😀😀😀😀😀😀😀😀😀😀😀😀😀😀😀😀😀😀😀😀😀😀😀😀😀😀😀😀😀😀😀😀😀😀😀😀😀😀😀😀😀😀😀a")" --job "Default" --benchmarkId 0 in 
// BeforeAnythingElse

// Benchmark Process Environment Information:
// Runtime=.NET Framework 4.8 (4.8.4300.0), X64 RyuJIT
// GC=Concurrent Workstation
// Job: DefaultJob

OverheadJitting  1: 1 op, 245600.00 ns, 245.6000 us/op
WorkloadJitting  1: 1 op, 210000.00 ns, 210.0000 us/op
...
```