@dnfadmin 2021-05-10
[![CLA assistant check](https://cla.dotnetfoundation.org/pull/badge/not_signed)](https://cla.dotnetfoundation.org/dotnet/BenchmarkDotNet?pullRequest=1705) <br/>Thank you for your submission, we really appreciate it. Like many open source projects, we ask that you sign our [Contributor License Agreement](https://cla.dotnetfoundation.org/dotnet/BenchmarkDotNet?pullRequest=1705) before we can accept your contribution.<br/><br/>:x: novak-as [sign now](https://cla.dotnetfoundation.org/dotnet/BenchmarkDotNet?pullRequest=1705)<br/><sub>You have signed the CLA already but the status is still pending? Let us [recheck](https://cla.dotnetfoundation.org/check/dotnet/BenchmarkDotNet?pullRequest=1705) it.</sub>