@novak-as 2021-05-11
```
CompletedWorkItemCounIsAccurate:
Assert.Equal() Failure
Expected: 1 (rounded from 1)
Actual:   1.001 (rounded from 1.0009765625)
```

I'm not sure if it's related to my recent change